requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
jinja2==3.1.3 
openai
python-dotenv
langchain
langchain_community